OU models now have "with.optimum" option.
Also much more extensive tests for OU cases.

The "with.optimum" argument represents a major interface change
for the OU models, but is important.  The default will be to fit
models like the fitContinuous models, where there is no optimum
parameter (and it will be fixed at something like its ML state).

For "pruning" based models only, the optimum can be treated as a
parameter, which is a slightly different version of OU.  It remains
unclear to me exactly how these are related.